// src/types/userType.ts
export type UserProfileType = {
  id: number
  firstName: string
  lastName: string
  email: string
  gender: string | null
  role: string
  phoneNumber: string
  address: string
  name: string
  companyId: number
  password: string
  updatedAt: string
  isDeleted: boolean
  isBlocked: string
  createdAt: string
}